+++ AstroDX12/Src/Rendering/RendererDX12.h
@@ -28,8 +28,8 @@ public:
 protected:
     virtual ComPtr<ID3D12Device>  GetDevice() const override { return m_device; };
 public:
-    virtual void CreateConstantBufferDescriptorHeaps(int16_t frameResourceCount, int32_t renderableObjectCount) override;
-    virtual void CreateConstantBufferView(D3D12_GPU_VIRTUAL_ADDRESS cbvGpuAddress, UINT cbvByteSize, int32_t handleOffset) override;
+    virtual void CreateConstantBufferDescriptorHeaps(size_t frameResourceCount, size_t renderableObjectCount) override;
+    virtual void CreateConstantBufferView(D3D12_GPU_VIRTUAL_ADDRESS cbvGpuAddress, UINT cbvByteSize, size_t handleOffset) override;
     virtual void CreateGraphicsPipelineState(
         ComPtr<ID3D12PipelineState>& pso,
         ComPtr<ID3D12RootSignature>& rootSignature,
@@ -39,7 +39,7 @@ public:
     virtual void BuildFrameResources(std::vector<std::unique_ptr<FrameResource>>& outFrameResourcesList, int frameResourcesCount, int objectCount) override;
     virtual UINT64 GetLastCompletedFence() override;
     virtual void WaitForFence(UINT64 fenceValue) override;
-    virtual void SetPassCBVOffset(int32_t offset) override;
+    virtual void SetPassCBVOffset(size_t offset) override;
     // IRenderer - END
 
 private:
@@ -66,7 +66,7 @@ private:
     static const uint8_t m_swapChainBufferCount = 2;
     int m_currentBackBuffer = 0;
     int m_currentFence = 0;
-    int32_t m_renderPassCBVOffset = 0;
+    size_t m_renderPassCBVOffset = 0;
 
     ComPtr<IDXGIFactory4> m_dxgiFactory;
     ComPtr<ID3D12Device> m_device;
